requirements.txt
fastapi==0.100.0
hypercorn==0.14.4
pydantic-settings>=2.0.0

# Geospatial
rasterio>=1.3.0
shapely>=2.0.0
pyproj>=3.6.0

# Data / ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0

# HTTP (SoilGrids REST fallback)
requests>=2.31.0
